requirements.txt
Flask==2.0.0
Pillow==8.2.0
requests==2.25.1
Werkzeug==2.0.1
gunicorn==20.1.0